feat(watchos): arm64_32 device support (Series 4-8 / SE)
Real pre-S9 Apple Watches use the arm64_32 architecture (64-bit ISA,
32-bit pointers). Perry never produced a working binary for them, and an
old note claimed it was architecturally impossible (NaN-boxing needs
64-bit usize). That's overstated: a 32-bit pointer fits trivially in the
48-bit NaN payload — only a few heap-range guard heuristics and literal
constants assumed a 64-bit address space.

Runtime (32-bit literal fixes; also benefit wasm32):
- array/from_concat: (1usize << 53) saturates via try_from on 32-bit
- box.rs / formatting.rs / value/dynamic_object.rs: 2^47/2^48 pointer-
  guard upper bounds compared in u64 so the literal isn't out-of-range
  for 32-bit usize (no-op on 32-bit — no addresses reach that high)
The heap-min *value* guards already pick the low (0x1000) floor on
target_os=watchos, so no value changes were needed there.

Compile driver (opt-in via PERRY_WATCHOS_ARM64_32, so the default
watchos=arm64/S9+ path is untouched):
- codegen emits the C entry via PERRY_ENTRY_SYMBOL (e.g. _perry_user_main)
  instead of renaming _main with rust-objcopy afterwards — objcopy's
  MachOWriter segfaults on arm64_32 Mach-O. The link-time objcopy rename
  is skipped on arm64_32 accordingly.
- helpers.rs / app_metadata.rs / link/mod.rs / platform_cmd.rs /
  bundle_apple.rs resolve the arm64_32-apple-watchos triple (codegen,
  auto-optimize runtime rebuild, native-lib build, swiftc, link) and a
  low MinimumOSVersion (PERRY_WATCHOS_MIN, default 11.0).

Validated end to end: built a real game (Bloom Jump) — perry-runtime +
perry-ui-watchos + engine native lib all compile for arm64_32-apple-
watchos; the game links to a Mach-O 'architecture: arm64_32, platform
WATCHOS, minos 11.0' with __perry_user_main and _main both defined and
no undefined refs. (Testable only on real pre-S9 hardware — the
simulator is arm64.)

Co-authored from a parked WIP exploring this target.

+++ crates/perry-runtime/src/box.rs
@@ -87,7 +87,7 @@ pub fn scan_box_roots_mut(visitor: &mut crate::gc::RuntimeRootVisitor<'_>) {
             // address (alloc gives 8-aligned pointers in user space)
             // matches `is_plausible_box_ptr` to keep this a no-op for
             // any pathological entry.
-            if addr >= 0x1000 && addr < 0x0001_0000_0000_0000 && addr % 8 == 0 {
+            if addr >= 0x1000 && (addr as u64) < 0x0001_0000_0000_0000 && addr % 8 == 0 {
                 unsafe {
                     visitor.visit_nanbox_f64_raw_slot(&raw mut (*ptr).value);
                 }
@@ -200,7 +200,7 @@ fn is_plausible_box_ptr(ptr: *mut Box) -> bool {
     if addr < 0x1000 {
         return false;
     }
-    if addr >= 0x0001_0000_0000_0000 {
+    if (addr as u64) >= 0x0001_0000_0000_0000 {
         return false;
     }
     if addr % std::mem::align_of::<Box>() != 0 {
